# tpl/tplimpl: Reimplement the ".Params tolower" template transformer