All `.Params` are stored lowercase, but it should work to access them `.Page.camelCase` etc. There was, however, some holes in the logic with the old transformer.

This commit fixes that by applying a blacklist instead of the old whitelist logic. `.Param` is a very distinct key. The original case will be kept in `.Data.Params.myParam`, but other than that it will be lowercased.

This implementation is also faster:

```bash
benchmark                                old ns/op     new ns/op     delta
BenchmarkTemplateParamsKeysToLower-4     15237         9581          -37.12%

benchmark                                old allocs     new allocs     delta
BenchmarkTemplateParamsKeysToLower-4     92             40             -56.52%

benchmark                                old bytes     new bytes     delta
BenchmarkTemplateParamsKeysToLower-4     2656          1360          -48.80%
```

Fixes #5068